feat(components-react/nextjs): add dynamic theme switcher | bh | #1185
