+++ packages/components-react/projects/nextjs/components/index.ts
@@ -1,5 +1,5 @@
 export * from './FooterPartials';
 export * from './HeaderPartials';
-export * from './Select';
+export * from './Selects';
 export * from './TableWithCaption';
 export * from './Toast';
